# [vt/sqlparser] Use strings.Builder in compliantName