It halves allocations.
I would like to use it for Myprintf also very much, but I definitely need more research :)