@LK4D4 2018-12-19
@sougou nope, nothing in `sqlparser` benchmarks shows any difference :)
But strings builder is basically `bytes.Buffer` which can be used only once because it uses unsafe trick to not allocate string on `String()`:
https://golang.org/src/strings/builder.go?s=1379:1412#L36
It was made for such cases.