@LK4D4 2018-12-19
@sougou 
I wrote
```
+
+func BenchmarkCompliantName(b *testing.B) {
+       testString := ".ab"
+       for i := 0; i < b.N; i++ {
+               _ = NewColIdent(testString).CompliantName()
+               _ = NewTableIdent(testString).CompliantName()
+       }
+}

```
Results are:
```
benchmark                    old ns/op     new ns/op     delta
BenchmarkCompliantName-8     231           75.8          -67.19%

benchmark                    old allocs     new allocs     delta
BenchmarkCompliantName-8     4              2              -50.00%

benchmark                    old bytes     new bytes     delta
BenchmarkCompliantName-8     230           16            -93.04%
```