@slanning 2018-12-24
Note that this requires go 1.10+
(1.9: "go/vt/sqlparser/ast.go:3694:10: undefined: strings.Builder";
https://golang.org/pkg/strings/#Builder )
bootstrap.sh only requires that `go version` doesn't fail, though some docs mention installing 1.11+